feat: visualize generated image's featuremap
