+++ models/encoder/filtering/channel_selection/select_all.py
